move member to pkg
